+++ packages/terra-application-links/src/terra-dev-site/doc/example/ApplicationTabsExample.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 /* eslint-disable import/no-extraneous-dependencies, import/no-unresolved, import/extensions */
 import { ApplicationTabs } from 'terra-application-links/lib/ApplicationLinks';
 import ApplicationLinkConfig from 'terra-application-links/lib/terra-dev-site/doc/common/ApplicationLinkConfig';
-import demoStyles from './demoStyles.scss';
+import demoStyles from './demoStyles.module.scss';
 /* eslint-enable import/no-extraneous-dependencies, import/no-unresolved, import/extensions */
 
 const ApplicationTabsExample = () => (