+++ packages/terra-embedded-content-consumer/src/terra-dev-site/provider/embedded-content-consumer/providers/CustomEventsProvider.provider.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Provider } from 'xfc';
 import classNames from 'classnames/bind';
 import ProviderTestTemplate from '../../../../EmbeddedContentProviderTestTemplate';
-import styles from './EmbeddedContentConsumerCommon.scss';
+import styles from './EmbeddedContentConsumerCommon.module.scss';
 
 const cx = classNames.bind(styles);