+++ packages/terra-hookshot/src/terra-dev-site/test/hookshot/HookshotCoordinates.test.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import classNames from 'classnames/bind';
 import Hookshot from '../../../Hookshot';
 import HookshotContent from '../common/HookshotContentTestTemplate';
-import styles from '../common/HookshotTestDocCommon.scss';
+import styles from '../common/HookshotTestDocCommon.module.scss';
 
 const cx = classNames.bind(styles);