+++ packages/terra-navigation-side-menu/src/terra-dev-site/doc/example/NavigationSideMenuExample.jsx
@@ -3,7 +3,7 @@ import classNames from 'classnames/bind';
 // eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions */
 import NavigationSideMenu from 'terra-navigation-side-menu/lib/NavigationSideMenu';
 
-import styles from './NavigationSideMenuExample.scss';
+import styles from './NavigationSideMenuExample.module.scss';
 
 const cx = classNames.bind(styles);